+++ src/assets/nacionalidades.json
@@ -3,10 +3,5 @@
       "userId": 1,
       "id": 1,
       "title": "Peru"
-    },
-    {
-      "userId": 1,
-      "id": 2,
-      "title": "Venezuela"
     }
   ]
